feat!: upgrade to svgo2.
The svg-to-vue api did not change, the function still returns a promise and
still accepts an object to set the svgo options and the path.

It is however still a breaking change because the format of the svgo options
object itself has changed between svgo1 an svgo2, so if the user passes
this object to the function they will need to change to the new svgo config
format otherwise svgo will throw an error.